--- app/cdap/components/Grid/index.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import GridHeaderCell from './GridHeaderCell';
-import Box from '@material-ui/core/Box';
-import GridKPICell from './GridKPICell';
-import GridTextCell from './GridTextCell';
-
-const GridView = () => {
-  return (
-    <Box>
-      {/* <GridHeaderCell /> */}
-      {/* <GridKPICell /> */}
-      {/* <GridTextCell /> */}
-    </Box>
-  );
-};
-
-export default GridView;
